+++ packages/cli/src/index.ts
@@ -3,7 +3,7 @@ import { program } from 'commander';
 import { init } from './commands/init';
 import { add } from './commands/add';
 
-program.version('0.1.5').description('CLI tool for managing components');
+program.version('0.1.7').description('CLI tool for managing components');
 
 program.command('init').description('Initialize the package').action(init);